♻️ Refactor: Dropdown 테스트 옵션 상수 분리
